Add files via upload
We perform energy analysis using 12 different building shapes. The buildings differ with respect to the 
glazing area, the glazing area distribution, and the orientation, amongst other parameters. We simulate 
various settings as functions of the afore-mentioned characteristics to obtain 768 building shapes. The 
dataset comprises 768 samples and 8 features, aiming to analyse the energy consumption patterns in 
buildings and to predict two real valued responses.